refactor: add FaucetDialog and useFaucetDialog
