+++ packages/app/src/systems/Faucet/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useCaptcha';
 export * from './useFaucet';
+export * from './useFaucetDialog';
